Run APK tests on the main thread.
Many tests rely on being run on the main thread. In this CL we post a delayed task to run tests on the main
thread and do not block in onCreate(). This can cause ANR if the main test activity receives a touch event, but this should not happen.
BUG=125059

TEST=


Review URL: https://chromiumcodereview.appspot.com/10663024

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 256a41cf5448f20d736a797075e9acafff4fa363