Split the power monitoring feature from SystemMonitor
The SystemMonitor is a mixed monitor which not only monitors
the power state changes but also the devices changes. This
patch is to separate the power monitor from SystemMonitor
as a new class PowerMonitor which is dedicated to monitor
power state.

The next step is to seek a opportunity to refactor SystemMonitor
as something like DeviceMonitor.

BUG=149059
TEST=base_unittests --gtest_filter=PowerMonitorTest.*

Review URL: https://chromiumcodereview.appspot.com/10959020

Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6c5905b763d18bf2d81e2d1f4aac7f16ae32f6ea