Get Win32K lockdown for renderers working again on Windows 8+.
This regressed recently with the change to query the device scale factor to see if it needs to be scaled back to 1.0 for
display scales equal to or under 125%. user32/gdi32 API's crash or fail with the Win32K lockdown process mitigation.

Fix is to ensure that the device scale factor is queried from the one set via the command line. The other change is to move
the code in ContentMainRunner::Initialize which sets the global device scale factor for the process to before the content client
initialization. This is because the ShellContentRendererClient in its ctor calls the ShouldUseDirectWrite function which relies on
the GetDPIScale function to decide whether or not directwrite should be enabled. This causes a DCHECK in content_browsertests.

To ensure that the win32k lockdown does not regress I added a simple content browsertest RendererWin32KLockdownNavigationTest which runs
this on Windows 8+.

The accessibility fixes are to get the browser tests to run
on Windows 8. There were special cases to always enable
accessibility on Windows 8 for historical reasons. These are
no longer necessary as we don't rely on accessibility to display the virtual keyboard.

BUG=417147

Review URL: https://codereview.chromium.org/600263002

Cr-Original-Commit-Position: refs/heads/master@{#296859}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b1f9313020d4c1e8619f21b5320a0255bf313b4b