[iossim] Update for Xcode 6.1 crash detection.
iossim relies on parsing the console log to detect crashes. As of Xcode
6.1, there is a new error message that indicates a crash, so this CL
updates iossim to signal that the simulated app crashed if any string in
a list of error strings is seen in the console log.

BUG=432213

Review URL: https://codereview.chromium.org/724063002

Cr-Original-Commit-Position: refs/heads/master@{#304032}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 10318845b6eaca2658d96df562c34579df5a08f2