Move the test template to //testing/test.gni (part 1)
This is the first CL to move the test template to //testing/test.gni

This CL is just creating an empty test.gni file. It is needed because
the test template is used in Blink, so we need to makes this change in
stages.

R=cjhopman@chromium.org

Review URL: https://codereview.chromium.org/850743003

Cr-Original-Commit-Position: refs/heads/master@{#311341}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c6c612cc34b18dbce1ba479419601744cec7c699