Roll DEPS to class-dump 3.5
Removes some build warnings.

Moves the class-dump repo to a github auto-mirror instead of a mirror
that needs to be manually updated.

This also requires that the class-dump.gyp file move from the
third-party repo into the main source tree. Note that the file
structure of the class-dump source changed so the .gyp had to be
updated. For comparison, the old .class-dump.gyp can be found at
https://chromium.googlesource.com/chromium/deps/class-dump/+/master/class-dump.gyp

BUG=318160

Review URL: https://codereview.chromium.org/1106993002

Cr-Original-Commit-Position: refs/heads/master@{#330798}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1c80791ce7c29a29fc816d89d1a5fa2ac6503b3e