Reland "Roll DEPS to class-dump 3.5"
Removes some build warnings.

Moves the class-dump repo to a github auto-mirror instead of a mirror
that needs to be manually updated.

This also requires that the class-dump.gyp file move from the
third-party repo into the main source tree. Note that the file
structure of the class-dump source changed so the .gyp had to be
updated. For comparison, the old .class-dump.gyp can be found at
https://chromium.googlesource.com/chromium/deps/class-dump/+/master/class-dump.gyp

BUG=318160, 417264

Review URL: https://codereview.chromium.org/1161813003

Cr-Original-Commit-Position: refs/heads/master@{#332485}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b61668a8e23fcdbe8fd1e851c836a1a4d45fb054