Disable a test on CFI buildbot that causes it to be red.
The test itself is fine, but it requires a bot clobber,
because gyp is missing a dependency on
tools/cfi/blacklist.txt. The plan is to remove the test from
the buildbot (this CL), and then add the dependency to gyp.

BUG=chromium:515347

Review URL: https://codereview.chromium.org/1269553006

Cr-Original-Commit-Position: refs/heads/master@{#341774}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: eb1c5795d835511a81fce1bcdd88e2dec098b636