Obtain host-log file-names from stdout of host-start process.
Recently, the location where me2me host logs are written changed on the Swarming bots (root-cause not yet clear). To be more resilient to such changes, to get the host log file-name we'll parse the output of the process that starts the host.

Also, in order to rule out replication delay as a cause of test failures, we'll check the JID used by a test and, if it is different from the JID logged by the host for that execution, sleep for 30-seconds and rerun the test if it failed.

BUG=

Review URL: https://codereview.chromium.org/1375373006

Cr-Original-Commit-Position: refs/heads/master@{#353409}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c989dcedc1a91b063bbd732c1e992d6c6181eb9c