Adjust OWNERS for //testing.
We used to have a wildcard OWNERS entry for //testing, and so
some of the sub-directories had "set noparent" directives.

The wildcard is now gone, so this CL removes the "set noparent"
and updates comments to be a bit clearer.

R=phajdan.jr@chromium.org
BUG=

Review URL: https://codereview.chromium.org/1415633004

Cr-Original-Commit-Position: refs/heads/master@{#358742}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2a7b6dee72d611cb9abac487ed24cf70f3045fc