Ensure remoting is built on the desktop GN bots.
Most of the remoting tests are not run on the bots, and so
we don't automatically get build coverage of the remoting targets,
and build failures may creep in.

This patch ensures that remoting_all is listed as an
additional_compile_target on the bots that aren't configured to
build 'gn_all', so that we can get as much build coverage as possible.

Also, remove the //testing/buildbot entry for 'Win GN', since that
bot doesn't exist any more.

R=sergeyu@chromium.org
BUG=530733, 512899

Review URL: https://codereview.chromium.org/1426393002

Cr-Original-Commit-Position: refs/heads/master@{#359252}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: af3e86deb9158be7b15024aff9b5c66b50a11ccf