Provide default implementation for WebContentsDelegate::OpenURLFromTab.
With OOPIF, when a frame in a background page is navigated cross-site
from the renderer, the cross-process transfer involves using
WebContentsDelegate::OpenURLFromTab to reissue the navigation, but the
delegate in this case is extensions::ExtensionHost, which doesn't
implement that method.  As a result, the default delegate
implementation is used, which just returns nullptr and drops the
request.

This CL provides a better default implementation for
WebContentsDelegate::OpenURLFromTab, borrowing it from
Shell::OpenURLFromTab.  This fixes the issue with background pages and
will hopefully prevent this type of issue from happening in other
contexts.

BUG=560510
TEST=ExtensionApiTest.(WindowOpenPopupIframe|ContentScriptExtensionProcess|DontInjectContentScriptsInBackgroundPages) with --isolate-extensions or --site-per-process

Review URL: https://codereview.chromium.org/1473423002

Cr-Original-Commit-Position: refs/heads/master@{#361724}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 35229c1ac03fd2707c8612973f6dd54a5150e9c4