Excluding SSLUITest.TestMarkNonSecureAs from Site Isolation FYI bot.
Excluding, because the test was very flaky on the Site Isolation Win bot.

The test was ok on the Site Isolation Linux bot, but I am excluding it
both on Linux and Win for consistency / to aid with maintaining the
exclusion lists.

BUG=417518
TBR=creis@chromium.org

Review URL: https://codereview.chromium.org/1491843005

Cr-Original-Commit-Position: refs/heads/master@{#362759}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e0e805b90abeac7aeca46ca8b7e5e9825bf0009e