Fix a bug where proxy cleanup during OOPIF navigations would detach frames. The fix involves reworking how proxy lifetimes are managed, so that FrameMsg_SwapOut is sent prior to the proxy shutdown IPCs. Add new tests of navigation vs. origin inheritance behavior, one of which (NavigateChildToAboutBlank) is what originally caught bug.
Prior to this change, the RFHM was responsible for meticulously calling
ShutdownProxiesIfLastActiveFrameInSiteInstance whenever an RFH was about to be
deleted or swapped out. In the new scheme, at the moment its
active_frame_count() hits zero, a SiteInstanceImpl will synchronously notify the
RFHMs that own a proxy in that site instance. The new behavior is achieved by
means of a new SiteInstanceImpl::Observer interface. Because each
RenderFrameHostManager observes all the SiteInstances that it created a proxy
for, it's no longer necessary to iterate over every
RenderWidgetHosts/RenderFrameHosts looking for proxies, so this is also more
efficient.

Importantly, the new scheme makes it possible to do proxy shutdown after
FrameMsg_SwapOut, which fixes the bug. In the old scheme,
ShutdownProxiesIfLastActiveFrameInSiteInstance had to be called while
active_frame_count was 1, which means it needed to occur before the SwapOut
operation. That ordering was fundamentally buggy: since the swapped-out frame's
parent was a proxy, cleaning up the proxy would delete the frame before
FrameMsg_SwapOut ever arrived, so there was no hope of FrameMsg_SwapOut being
received, or a SwapOut ACK ever being returned.

In addition to the ActiveFrameCountIsZero() observer notification, there is also
a RenderProcessGone() notification from SiteInstanceImpl to the
RenderFrameHostManager. Since SiteInstanceImpl already implemented
RenderProcessHostObserver, it's unnecessary for RenderFrameHostManager to be
have a RenderProcessHostObserver too, so we replace the elaborate
RenderFrameProxyHostMap with a simple container.

Note that even with 150 lines of new tests, this CL is a net deletion of code.

BUG=571637, 564292, 474246

Review URL: https://codereview.chromium.org/1485193004

Cr-Original-Commit-Position: refs/heads/master@{#368984}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9f34e289e4ecf80a72c06ec6ed0d38f6daca23d8