Disable angle_deqp_gles2_tests on Linux.
The Chromium recipe uses GN on Linux and this target is currently
commented out in the ANGLE workspace. Unfortunately this makes the Linux
dEQP bot a no-op for the moment, but this must be done to allow the
builder to work.

BUG=542370
TBR=dpranke@chromium.org
NOTRY=true

Review URL: https://codereview.chromium.org/1648753004

Cr-Original-Commit-Position: refs/heads/master@{#372313}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6f24e91373eda309e1df55f69f2e5df86f99d485