Rework the way args are passed to run_under_valgrind.py slightly.
This is a preparatory CL reworking the way we pass arguments
to the run_under_valgrind.py script on the bots, to make things
a little more transparent.

This CL should have no effect on the bots, but makes the follow-on
CL to enable sharding of the unit_tests a little smaller and easier
to review.

R=bruening@chromium.org
BUG=572694

Review URL: https://codereview.chromium.org/1691863003

Cr-Original-Commit-Position: refs/heads/master@{#375051}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bf38b83f7f8b2c74b6cba6060a375b2aab5c0ea6