Run mojo_apptests on debug Linux as well.
Any test which runs on the main waterfall should run on debug and release. Otherwise a dcheck could be hit which wouldn't show up on the main waterfall, but would fail on the trybots which use release-with-asserts.

BUG=478244

Review URL: https://codereview.chromium.org/1705883002

Cr-Original-Commit-Position: refs/heads/master@{#376028}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a7cef43c8ed1c379bd2d6bb5fffd3a7ff2a38fb0