Removing the event server and putting in place a more versatile comm server.
The server routes requests to specific handlers based on the first part of
the path. For example URL/messages/... will be routed to the MessageHandler.

The message handler is a much improved version of the event handler. The handler
can be used as an event server by ommiting a body and using the return code as
the true/false value depending on if its been set before. However it can be used
as an inter-test component messaging system when uploading a body.

BUG=

Review URL: https://codereview.chromium.org/1714983002

Cr-Original-Commit-Position: refs/heads/master@{#376775}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 85d400e763d08c5be08429bfdef8b5d3ceff3ad7