PlzNavigate: notify a WebUI of RVH creation when it is commited at creation
This CL makes it so that a newly created WebUI is notified that a
RenderViewHost was created for it when it is committed immediately after being
created (when the current RenderFrameHost is not live). This was not previously
the case, as the code tried to get the navigating WebUI, defined as the WebUI
of the speculative RenderFrameHost or the pending WebUI of the current
RenderFrameHost. In the case where the speculative RenderFrameHost is created,
the new WebUI becomes the current WebUI of the current RenderFrameHost, and
GetNavigatingWebUI() returns nullptr.

BUG=475027

Review URL: https://codereview.chromium.org/1733463004

Cr-Original-Commit-Position: refs/heads/master@{#378430}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 3b14d074ebf91788a01ee3b0544b66e1f65f1bca