Enable DRP config service by default
We have been running this experiment at 50% stable, and are
now ready to move to 100%.

This CL does not remove the code that will now be deprecated
because of DRP config being enabled 100% of time. We should
remove that code at some point in future.

This is reland of
https://codereview.chromium.org/1830343002/ which had to be
reverted because of failing tests.

BUG=597768
TBR=holte@chromium.org

Review URL: https://codereview.chromium.org/1845773002

Cr-Original-Commit-Position: refs/heads/master@{#384319}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 57d8bc8d65f8bc73ce8c74e7b092c19002ca5c3a