Do not reset navigation state when BeforeUnload is cancelled by a commit
It is possible for a cross-site navigation to commit while a BeforeUnload
confirmation dialog is showing. This will destroy dialog and call
WebContentsImpl::OnDialogClosed. This function should not reset the navigatin
state if the RenderFrameHost that was showing the dialog is no longer the
current RenderFrameHost.

BUG=589365

Review URL: https://codereview.chromium.org/1825523002

Cr-Original-Commit-Position: refs/heads/master@{#384674}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bc6f4150a3eef02b926611404e7cdd44c9d67635