Add functionality to the install_static library to tokenize strings and compare versions.
This is required for the ongoing work to add crashpad integration in
chrome_elf and remove the crashpad integration from chrome.exe

As chrome_elf loads very early in the process, we cannot depend on
functionality from base which pulls in additional dependencies from
user32, etc.

I also added a target called install_static_unittests which contains tests
for the new functionality in install_static. The MatchPattern test has
been moved from chrome_elf to this executable.

The utf8_to_utf16 and utf16_to_utf8 functions in install_static are now public functions as they are
needed by crashpad.

The other changes in this patch are for getting the install_static
executable to run on the windows bots.

BUG=604923
TBR=dpranke

Review-Url: https://codereview.chromium.org/2017853002
Cr-Original-Commit-Position: refs/heads/master@{#396612}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f2e54a999d4bd88062e27c61f995bac80cc025b1