Prevent renderer kills for in-page navigations on subframes.
This switches HistoryController and HistoryEntry to look at the last
committed WebHistoryItem for a frame rather than assuming that the
whole HistoryEntry accurately reflects the state of the page.  The
previous logic led to NC_IN_PAGE_NAVIGATION renderer kills when going
back cross-site in ways that looked in-page to the HistoryEntry.

This CL does not yet fix the kills that happen in FrameNavigationEntry
modes, which are covered by the BackTwiceToIframeWithContent test.
Those will be fixed in a follow-up CL.

BUG=612713
TEST=Fewer NC_IN_PAGE_NAVIGATION kills.
CQ_INCLUDE_TRYBOTS=tryserver.chromium.linux:linux_site_isolation

Review-Url: https://codereview.chromium.org/2023013002
Cr-Original-Commit-Position: refs/heads/master@{#397832}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 225a74338e488576c405e18610298c39244b85f5