PlzNavigate: fix RenderViewImplTest.OnNavigationHttpPost
This CL fixes RenderViewImplTest.OnNavigationHttpPost when PlzNavigate is
enabled by having RenderFrameImpl::NavigateInternal chose to use LoadDataURL
instead of WebLocalFrame::load in the same cases as what the current
architecture is doing.

It also ensures that the WebURLLoader will not replace the ResourceResponse it
receives with one provided by the browser when loading a data url. In this
case, the browser has no response to provide. so we should use the one
generated by the HandleDataURL task.

BUG=475027

Review-Url: https://codereview.chromium.org/2037733002
Cr-Original-Commit-Position: refs/heads/master@{#398008}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 269f92ce7d43d77d11674ea7898b6d826eb579d5