Copy "iossim" from host toolchain instead of creating a symlink.
Chrome on iOS downstream bots package build output to ship it to the
swarming bots and does not follow symlinks while packing. Change the
"iossim" target to copy the output from the host toolchain instead of
creating a symlink to fix the bot (as we want them to work the same
with gyp and gn and the bot).

BUG=621708

Review-Url: https://codereview.chromium.org/2082113002
Cr-Original-Commit-Position: refs/heads/master@{#401112}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9fc0d12586615aa807b09bdf260f0fec24f2eec7