Add a fuzzer for srtp_unprotect in libsrtp.
This is called on any SRTP packet received, and decrypts it in
place. For now the fuzzer uses a hard-coded key and SSRC, but future
modifications could generalise this in a number of ways.

It does read more than one packet, since there could be bugs lurking
in the state maintained by the parser (for example, the ID of each
packet is stored for replay protection).

Review-Url: https://codereview.chromium.org/2123553002
Cr-Original-Commit-Position: refs/heads/master@{#403893}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 346b92100d3ebfe43ec8c226d8bf0fa226c22527