Make sure a test cannot be listed twice for one builder.
This check is now made at presubmit time, preventing hard to debug failure on
the infrastructure.

The failure mode is not obvious, the recipe throw an exception at trigger time
and this is hard to diagnose.

This fixes the problem before this is even sent to run as a try job.

Tested manually.

R=thakis@chromium.org
BUG=626766

Review-Url: https://codereview.chromium.org/2151503003
Cr-Original-Commit-Position: refs/heads/master@{#405194}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1e0384f82f2670e949e59556925978118246be92