Add msan=1 to GYP_DEFINES in libfuzzer+msan example.
Update the documentation to add msan=1 to the GYP_DEFINES in order to have the
pre-build binarys downloaded correctly.

Without msan=1 you get an error from the download_binaries script:

Exception: Prebuilt instrumented libraries not available for your configuration.

The error happens because there is a regex looking for msan=1 in the defines
and erroring otherwise.

Review-Url: https://codereview.chromium.org/2180963006
Cr-Original-Commit-Position: refs/heads/master@{#407819}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b8d922a9dee1a3c27a0569a0d275dad983c4bbc3