Run ShortcutHelper::AddToLauncherInBackgroundWithSkBitmap() on Worker thread instead of IO thread
This CL runs ShortcutHelper::AddToLauncherInBackgroundWithSkBitmap() on a
worker thread instead of the IO thread. According to
https://www.chromium.org/developers/design-documents/threading expensive tasks
should not be done on the IO thread. This CL also changes
ShortcutHelper::AddToLauncherInBackgroundWithSkBitmap() so that only the
expensive parts are run on the worker thread instead of the entire process
started by ShortcutHelper::AddToLauncherInBackgroundWithSkBitmap() running on
the worker thread.

BUG=635347

Review-Url: https://codereview.chromium.org/2228273002
Cr-Original-Commit-Position: refs/heads/master@{#411743}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c67e6ac6fb0cc0b0ba6582ea1a9e8e62caf9748a