[iossim] Work around pipe hang on some macs.
In the rest of the this file we read from the pipe after -waitUntilExit
(We normally wrap -launch and -waitUntilExit in one -run method).  However,
on some swarming slaves this led to a hang on simctl's pipe.  Since the
output of simctl is so instant, reading it before exit seems to work, and
seems to avoid the hang.

BUG=
NOTRY=true

Review-Url: https://codereview.chromium.org/2240373002
Cr-Original-Commit-Position: refs/heads/master@{#411842}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7768f4337363cc5fb41ba14297b4bcff232ccd91