Change variations service domain for Android Chrome.
The new domain is clientservices.googleapis.com.

Also, removes testing config for VarationsServiceControl
30m interval trial, since that's the client default.
Updates the parameter code for the fetch interval to fix
the typo in the trial name.

Note: Should not be submitted until we verify that the new
URL is working.

BUG=635680

Review-Url: https://codereview.chromium.org/2223263002
Cr-Original-Commit-Position: refs/heads/master@{#415750}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1078363787edb725a21dd94993a1b8fdacd346cb