bluetooth: web: Update Blacklist field trial with Cypress & FIDO services.
In preparation for deploying the Web Bluetooth Blacklist update via
field trial parameter this patch adds the payload to the
Field Trial Testing Config.

To support a browser test for this field trial parameter,
a test specific UUID is also added to the configs.

To enable readable comments in the json, the PRESUBMIT is
modified to accept comment keys.

BUG=570121, 643013

Review-Url: https://codereview.chromium.org/2306473002
Cr-Original-Commit-Position: refs/heads/master@{#416332}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 45e66c790f6b93dec90df68aaaf826c09ae114a1