Process DisplaySnapshots in PlatformScreen.
Add logic to process DisplaySnapshots from the DisplayConfigurator.
Process the contents of the snapshots to find new, removed or modified
displays.

Add unit tests to test that the correct calls are made to the
PlatformScreenDelegate in response to receiving an updated list of
display snapshots.

BUG=641012

Review-Url: https://codereview.chromium.org/2297743002
Cr-Original-Commit-Position: refs/heads/master@{#416657}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 80f206eec2f4ba6ec5485d6a40ac5a9af968ab38