Fix swarming for content_browsertests and browser_tests on PlzNavigate FYI bot.
The problem was that the path to the filter is different depending on whether the test is swarmed or not. Also make sure to send the filter file for browser_tests as part of the isolate data.

BUG=645983
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/2335333003 .

Cr-Original-Commit-Position: refs/heads/master@{#418421}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a8a834050404cd8bda8d54addfd146320d0981e2