Make sure changes to //third_party/typ cause tests to be run.
Given that typ is part of the testing infrastructure, we should
make sure that when typ is rolled, the tests are run. Ideally
this would be done by having the correct data deps listed in GN,
but we don't have a strict mapping of test steps to build targets yet
(e.g., there's no build target for webkit_python_tests) and so
a blanket whitelist is a bit safer.

R=nednguyen@google.com
BUG=

Review-Url: https://chromiumcodereview.appspot.com/2417453002
Cr-Original-Commit-Position: refs/heads/master@{#426276}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9a35e8d6fb4a63a96d8136c0f00916a6d560fd47